Bindings/sysworxx-io-js/Cargo.toml: Link libsysworxx_io.so statically
Having it as a build-dependencies adds dependencies to libiio /
libsensors to the host system which is not what we want to
achieve here.

Adding it as a normal dependency and specify dynamic linking via
`build.rs` will not work as well. (like working for libctr700drv.so)
When inspecting the generated .so file one can see that libsysworxx_io
is still statically linked. This maybe related to the use of neon, not
sure about this.

Since we deploy Node-RED nodes without this library we should be fine for
now by statically linking.

Signed-off-by: Andreas Dinter <andreas.dinter@systec-electronic.com>